requirements.txt
flask==3.0.3
qrcode==7.4.2
pillow==10.4.0
openpyxl==3.1.5
requests==2.32.3
